Move internal/nlwrap to daemon/libnetwork/nlwrap
