+++ daemon/libnetwork/drivers/bridge/interface_linux.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/containerd/log"
 	"github.com/docker/docker/daemon/libnetwork/internal/netiputil"
+	"github.com/docker/docker/daemon/libnetwork/nlwrap"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/internal/nlwrap"
 	"github.com/vishvananda/netlink"
 )
