+++ daemon/libnetwork/drivers/overlay/ov_utils.go
@@ -12,8 +12,8 @@ import (
 	"github.com/containerd/log"
 	"github.com/docker/docker/daemon/libnetwork/drivers/overlay/overlayutils"
 	"github.com/docker/docker/daemon/libnetwork/netutils"
+	"github.com/docker/docker/daemon/libnetwork/nlwrap"
 	"github.com/docker/docker/daemon/libnetwork/ns"
-	"github.com/docker/docker/internal/nlwrap"
 	"github.com/vishvananda/netlink"
 	"github.com/vishvananda/netns"
 )
