+++ daemon/libnetwork/iptables/conntrack.go
@@ -9,8 +9,8 @@ import (
 	"syscall"
 
 	"github.com/containerd/log"
+	"github.com/docker/docker/daemon/libnetwork/nlwrap"
 	"github.com/docker/docker/daemon/libnetwork/types"
-	"github.com/docker/docker/internal/nlwrap"
 	"github.com/vishvananda/netlink"
 )
