+++ daemon/libnetwork/libnetwork_linux_test.go
@@ -25,10 +25,10 @@ import (
 	"github.com/docker/docker/daemon/libnetwork/ipams/null"
 	"github.com/docker/docker/daemon/libnetwork/ipamutils"
 	"github.com/docker/docker/daemon/libnetwork/netlabel"
+	"github.com/docker/docker/daemon/libnetwork/nlwrap"
 	"github.com/docker/docker/daemon/libnetwork/options"
 	"github.com/docker/docker/daemon/libnetwork/osl"
 	"github.com/docker/docker/daemon/libnetwork/types"
-	"github.com/docker/docker/internal/nlwrap"
 	"github.com/docker/docker/internal/testutils/netnsutils"
 	"github.com/docker/docker/pkg/plugins"
 	"github.com/moby/sys/reexec"
