+++ daemon/libnetwork/osl/sandbox_linux_test.go
@@ -12,9 +12,9 @@ import (
 	"syscall"
 	"testing"
 
+	"github.com/docker/docker/daemon/libnetwork/nlwrap"
 	"github.com/docker/docker/daemon/libnetwork/ns"
 	"github.com/docker/docker/daemon/libnetwork/types"
-	"github.com/docker/docker/internal/nlwrap"
 	"github.com/docker/docker/internal/testutils/netnsutils"
 	"github.com/vishvananda/netlink"
 	"github.com/vishvananda/netlink/nl"
