+++ integration-cli/docker_cli_network_unix_test.go
@@ -19,9 +19,9 @@ import (
 	"github.com/docker/docker/daemon/libnetwork/ipamapi"
 	remoteipam "github.com/docker/docker/daemon/libnetwork/ipams/remote/api"
 	"github.com/docker/docker/daemon/libnetwork/netlabel"
+	"github.com/docker/docker/daemon/libnetwork/nlwrap"
 	"github.com/docker/docker/integration-cli/cli"
 	"github.com/docker/docker/integration-cli/daemon"
-	"github.com/docker/docker/internal/nlwrap"
 	"github.com/docker/docker/pkg/plugins"
 	"github.com/docker/docker/pkg/stringid"
 	"github.com/docker/docker/testutil"
