+++ integration-cli/docker_cli_swarm_test.go
@@ -22,10 +22,10 @@ import (
 	"github.com/docker/docker/daemon/libnetwork/driverapi"
 	"github.com/docker/docker/daemon/libnetwork/ipamapi"
 	remoteipam "github.com/docker/docker/daemon/libnetwork/ipams/remote/api"
+	"github.com/docker/docker/daemon/libnetwork/nlwrap"
 	"github.com/docker/docker/integration-cli/checker"
 	"github.com/docker/docker/integration-cli/cli"
 	"github.com/docker/docker/integration-cli/daemon"
-	"github.com/docker/docker/internal/nlwrap"
 	"github.com/docker/docker/pkg/plugins"
 	"github.com/docker/docker/testutil"
 	testdaemon "github.com/docker/docker/testutil/daemon"
