+++ integration/daemon/daemon_linux_test.go
@@ -7,7 +7,7 @@ import (
 	"slices"
 	"testing"
 
-	"github.com/docker/docker/internal/nlwrap"
+	"github.com/docker/docker/daemon/libnetwork/nlwrap"
 	"github.com/docker/docker/internal/testutils/networking"
 	"github.com/docker/docker/testutil"
 	"github.com/docker/docker/testutil/daemon"
