+++ testutil/environment/protect_linux.go
@@ -11,7 +11,7 @@ import (
 	"testing"
 
 	"github.com/docker/docker/daemon/libnetwork/drivers/bridge"
-	"github.com/docker/docker/internal/nlwrap"
+	"github.com/docker/docker/daemon/libnetwork/nlwrap"
 	"github.com/vishvananda/netlink"
 	"gotest.tools/v3/assert"
 )
